+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+- Add `CAS\CasService` service, accessible through `cas` view helper and `cas`
+  controller plugin. It provides a new method `isCasUser` which checks if a
+  user used CAS to log in.
+- The service also provides `loginUrl` which does the same as `casLoginUrl`
+  view helper, which is now deprecated.
+
 ## [0.7.0] - 2026-04-13
 
 - Allow to assign groups to created users, if module Group is enabled